added the socket api

+++ routers/AuthRoutes.js
@@ -4,6 +4,8 @@ const { User } = require("../models/User");
 const { generateToken } = require("../config/JwtConfig");
 const { Device } = require("../models/Devices");
 const { Business } = require("../models/Business");
+const { default: axios } = require("axios");
+const { SOCKET_API } = require("../config/config");
 const router = express.Router();
 
 router.post("/login", loginWithPhone);
@@ -43,11 +45,23 @@ router.post("/remove-device-login", async (req, res) => {
 
     const { businessId } = user.dataValues;
 
-    await Device.destroy({
+    const devices = await Device.findAll({
       where: {
         businessId: businessId,
       },
     });
+
+    if (devices.length > 0) {
+      await Device.destroy({
+        where: {
+          businessId: businessId,
+        },
+      });
+      await axios.post(SOCKET_API + "/logout", {
+        roomId: businessId,
+        devices: devices,
+      });
+    }
 
     await Device.create({
       fcmToken,
@@ -70,7 +84,7 @@ router.post("/remove-device-login", async (req, res) => {
       user,
     });
   } catch (error) {
-    return res.json({ message: "Something went wrong", status: false,error });
+    return res.json({ message: "Something went wrong", status: false, error });
   }
 });
